Basic implementation of state pattern with unique pointer
Just a very basic implementation of the state pattern with a unique pointer for the abstract state. While make unique can't be used with abstract type a unique pointer type can.
Therefore, to determine which child class to make unique we use an enum to identify the class type, with this we can then make the unique pointer of the abstract type be an instance of a non-abstract child class through make unique